+++ arena_tactics/autoloads/settings/setting.gd
@@ -2,7 +2,6 @@ extends Node
 class_name Setting
 
 # Class for defining a single setting.
-@export_group("Mandatory")
 @export var key: String
 @export_multiline var tooltip: String
 var value: Variant
@@ -12,21 +11,28 @@ enum SETTING_TYPE {
 	BOOLEAN,
 	RANGE,
 	OPTIONS,
+	TEXT,
+	NUMBER
 }
 
 @export var type: SETTING_TYPE = SETTING_TYPE.BOOLEAN
 
-@export_group("Only when type is range")
-@export var min_value: int = 0
-@export var max_value: int = 100
-@export var step: float = 1
+@export_group("Range Type")
+@export var min_value_range: int = 0
+@export var max_value_range: int = 100
+@export var step_range: float = 1
 
-@export_group("Only when type is option")
+@export_group("Option Type")
 @export var possible_values: Array:
 	set(val):
 		possible_values = val
 
 @export var possible_values_strings: Array = []
+
+@export_group("Number Type")
+@export var min_value_nbr: int = 0
+@export var max_value_nbr: int = 100
+@export var step_nbr: float = 1
 
 func clear() -> void:
 	value = null